Restructure, python3 rewrite, removed bootstrapping
All required files now reside in the subdirectory cmus-osx.
Installation is easier this way.

Refactored and rewrote most code and added python3 support in the process.
python2.7 is not supported anymore. Most object orientation was also ripped out, as it was unnecessary for one instance during runtime and slow.

Instead of starting cmus from a python script the cmus rc is utilised to start the media-key handler.

More changes to follow python coding conventions more closely.